[agent] split list_tui.go by concern
